+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added 
+- [useScript](docs/composable/web/script) - appends `script` to the document header
+
+
 ## 0.2.0
 ------------------------------
 *2019-11-03*